backend/requirements.txt
fastapi
uvicorn
numpy
requests
python-dotenv
langchain
langchain-community
langgraph
langchain-groq
pinecone
huggingface_hub
langchain-pinecone
# # python-3.10.16   
# conda install numpy
# sentence-transformers


